extract audio code in audio utils. Update README
