+++ README.md
@@ -213,14 +213,12 @@ sentences = ["This is awesome.", "I am bored."]
 
 url = "https://bigsoundbank.com/UPLOAD/wav/2380.wav"
 raw_bytes = requests.get(url, stream=True).content
-data, samplerate = sf.read(io.BytesIO(raw_bytes))
 
-audios = [data]
+audios = [raw_bytes]
 engine_args = EngineArgs(
     model_name_or_path = "laion/clap-htsat-unfused",
     dtype="float32", 
     engine="torch"
-
 )
 array = AsyncEngineArray.from_args([engine_args])
 
@@ -232,6 +230,7 @@ async def embed(engine: AsyncEmbeddingEngine):
 
 asyncio.run(embed(array["laion/clap-htsat-unfused"]))
 ```
+*Note:* Audio Files need to have a sample rate of 48kHz.
 
 Example models:
 - [Clap Models from LAION](https://huggingface.co/collections/laion/clap-contrastive-language-audio-pretraining-65415c0b18373b607262a490)
